+++ spec/rspec/matchers/built_in/change_spec.rb
@@ -55,6 +55,15 @@ describe "expect { ... }.to change(actual, message)" do
     end
   end
 
+  context "with an IO stream" do
+    it "fails when the stream does not change" do
+      expect {
+        k = STDOUT
+        expect { }.to change { k }
+      }.to fail_with(/expected result to have changed/)
+    end
+  end
+
   context "with nil value" do
     before(:each) do
       @instance = SomethingExpected.new
@@ -212,6 +221,13 @@ describe "expect { ... }.to change { block }" do
 
   it "provides a #description" do
     expect(change { @instance.some_value }.description).to eq "change result"
+  end
+
+  context "with an IO stream" do
+    it "passes when the stream does not change" do
+      k = STDOUT
+      expect { }.not_to change { k }
+    end
   end
 end